pyproject.toml
[tool.poetry]
name = "hsi-penguinid"
version = "0.1.0"
description = ""
authors = ["dekkaiinu <24amj29@ms.dendai.ac.jp>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
torch = "2.3.0"
torchvision = "0.18.0"
omegaconf = "^2.3.0"
hydra-core = "^1.3.2"
tqdm = "^4.66.2"
opencv-python = "^4.9.0.80"
scikit-learn = "^1.4.2"
hsitools = {git = "https://github.com/dekkaiinu/hsitools"}
datasets = "^2.19.1"
h5py = "^3.11.0"
notebook = "^7.2.1"
yolov5 = "^7.0.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
